refactor: 상품재고에 대한 입력 패키지로 묶기
