+++ src/main/java/vendingmachine/dto/request/ItemPurchaseRequest.java
@@ -1,0 +1,2 @@
+package vendingmachine.dto.response;public class ItemPurchaseRequest {
+}
